requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.9
Werkzeug==0.14.1
